Update base.php
Ran into a 'Unable to open ...' error using the Template render method that was not generated using the View render method. There appeared to be a small inconsistency in parsing the UI global. So i added the small addition .';./' that was present on the View method, now they're both working fine.